requirements.txt
Django==6.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
psycopg2==2.9.11
dj-database-url==3.0.1
whitenoise==6.11.0
django-cors-headers==4.3.0
uvicorn ==0.40.0
